+++ src/internal.ts
@@ -1,4 +1,4 @@
-export * as product from "./product/internals";
+export * as product from "./product/internal";
 export * as parsing from "./parsing";
 export * as pdf from "./pdf";
 export * as geometry from "./geometry";
