+++ src/product/eu/internal.ts
@@ -1,0 +1,1 @@
+export * as licensePlate from "./licensePlate/internal";
