--- src/product/eu/internals.ts
@@ -1,1 +1,0 @@
-export * as licensePlate from "./licensePlate/internals";
