+++ src/product/fr/bankAccountDetails/internal.ts
