+++ src/product/fr/idCard/internal.ts
