+++ src/product/fr/internal.ts
@@ -1,0 +1,3 @@
+export * as bankAccountDetails from "./bankAccountDetails/internal";
+export * as carteVitale from "./carteVitale/internal";
+export * as idCard from "./idCard/internal";
