--- src/product/fr/internals.ts
@@ -1,3 +1,0 @@
-export * as bankAccountDetails from "./bankAccountDetails/internals";
-export * as carteVitale from "./carteVitale/internals";
-export * as idCard from "./idCard/internals";
