--- src/product/internals.ts
@@ -1,11 +1,0 @@
-export * as cropper from "./cropper/internals";
-export * as custom from "./custom/internals";
-export * as eu from "./eu/internals";
-export * as financialDocument from "./financialDocument/internals";
-export * as fr from "./fr/internals";
-export * as invoice from "./invoice/internals";
-export * as invoiceSplitter from "./invoiceSplitter/internals";
-export * as passport from "./passport/internals";
-export * as proofOfAddress from "./proofOfAddress/internals";
-export * as receipt from "./receipt/internals";
-export * as us from "./us/internals";
