+++ src/product/receipt/internal.ts
