+++ src/product/us/internal.ts
@@ -1,0 +1,2 @@
+export * as bankCheck from "./bankCheck/internal";
+export * as driverLicense from "./driverLicense/internal";
