--- src/product/us/internals.ts
@@ -1,2 +1,0 @@
-export * as bankCheck from "./bankCheck/internals";
-export * as driverLicense from "./driverLicense/internals";
